+++ components/FlexibleHeader/src/MDCFlexibleHeaderContainerViewController.m
@@ -1,3 +1,19 @@
+/*
+ Copyright 2015-present Google Inc. All Rights Reserved.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ */
+
 #if !defined(__has_feature) || !__has_feature(objc_arc)
 #error "This file requires ARC support."
 #endif
@@ -7,15 +23,21 @@
 #import "MDCFlexibleHeaderView.h"
 #import "MDCFlexibleHeaderViewController.h"
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
 @interface MDCFlexibleHeaderContainerViewController () <MDCFlexibleHeaderParentViewController>
 @end
+#pragma clang diagnostic pop
 
 @implementation MDCFlexibleHeaderContainerViewController
 
 - (instancetype)initWithContentViewController:(UIViewController *)contentViewController {
   self = [super initWithNibName:nil bundle:nil];
   if (self) {
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     [MDCFlexibleHeaderViewController addToParent:self];
+#pragma clang diagnostic pop
 
     self.contentViewController = contentViewController;
   }
@@ -36,7 +58,10 @@
   [self.view addSubview:self.contentViewController.view];
   [self.contentViewController didMoveToParentViewController:self];
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
   [self.headerViewController addFlexibleHeaderViewToParentViewControllerView];
+#pragma clang diagnostic pop
 }
 
 - (BOOL)prefersStatusBarHidden {