+++ demos/Shrine/Shrine/Shrine-Bridging-Header.h
@@ -1,0 +1,4 @@
+#import "MaterialButtons.h"
+#import "MaterialFlexibleHeader.h"
+#import "MaterialShadowLayer.h"
+#import "RemoteImageService.h"